fix(menuSelection): area selection behavior updated
Allows when a new area is selected, if there is a predicted layer selected it
will be deselected and the state for predicted layer updated.
As the selection of formulas and predicted layer depends on the state, it does
not make sense to have an update for the predicted layer depending on the
query params.